1.修复大量在订单，seller，商家密码业务方面的bug

+++ src/main/java/com/threeDays/dao/OrderMapper.java
@@ -6,7 +6,7 @@ import java.math.BigInteger;
 
 public interface OrderMapper {
     Order findOrderById(BigInteger order_id);//通过id查找order
-    BigInteger insertOrder(Order order);//插入新的order，order_id可以不用手动初始化,返回order_id
+    int insertOrder(Order order);//插入新的order，order_id可以不用手动初始化,返回order_id
     int updateOrder(Order order);//更新Order，order参数中必须手动初始化order_id，成功1，失败0
     int insertComment(BigInteger order_id,String commment);//插入买家评价，成功1，失败0
     int insertReply(BigInteger order_id,String reply);//插入卖家回复，成功1，失败0
